requirements.txt
asgiref==3.5.2
astroid==2.11.5
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
defusedxml==0.7.1
dill==0.3.5.1
Django==4.0.5
django-allauth==0.50.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
djlint==1.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
pathspec==0.9.0
pep8==1.7.1
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.0
python-dateutil==2.8.2
python3-openid==3.2.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.1
zipp==3.8.0
